Create README.md
The operating principle of this lab is based on a 4-bit adder. The inputs of this system are 8 switches (2 x 4 bits), and the outputs are LEDs and 4 7-segment displays. The word to be displayed is as follows: "S = XX". XX represents the decimal sum. The reset function allows the resetting of the processes of this adder.